+++ src/client/hooks/useHealth.ts
@@ -1,24 +1,44 @@
 // eslint-disable-next-line import/no-cycle
-import { useMemo } from 'react';
+
+import React from 'react';
 
 import client from '@src/client';
+import useRecursiveTimeout from '@src/hooks/useRecursiveTimeout';
+import { Health } from '@src/client/interfaces';
 
 export const useRpcBackup = () => {
     const { blockdiff } = useHealth();
+
+    console.log({ blockdiff });
     return blockdiff > 5;
+};
+
+const ok = (abc: Health): abc is Required<Health> => {
+    return (
+        (!!abc.lastBlockRpc || abc.lastBlockRpc === 0) &&
+        (!!abc.lastBlockGraph || abc.lastBlockGraph === 0)
+    );
 };
 
 export const useHealth = () => {
     const health = client.live.health();
 
-    const blockdiff = useMemo(() => {
-        return (
-            ((health.lastBlockRpc || health.lastBlockRpc === 0) &&
-                (health.lastBlockGraph || health.lastBlockGraph === 0) &&
-                Math.abs(health.lastBlockGraph - health.lastBlockRpc)) ||
-            Number(0)
-        );
-    }, [health]);
+    const [blockdiff, setBlockDiff] = React.useState(0);
 
-    return { blockdiff };
+    useRecursiveTimeout(
+        React.useCallback(() => {
+            setBlockDiff(
+                (ok(health) && Math.abs(health.lastBlockGraph - health.lastBlockRpc)) || Number(0),
+            );
+        }, [health]),
+        5000,
+    );
+
+    const graphProblem = React.useMemo(() => {
+        if (!ok(health)) return false;
+
+        return blockdiff > 2 && health.lastBlockGraph < health.lastBlockRpc;
+    }, [health, blockdiff]);
+
+    return { blockdiff, graphProblem };
 };
