+++ src/client/subscriptions/useLiveItem.ts
@@ -1,22 +1,15 @@
 import { useWatchLiveItemSubscription } from '@src/gql/types.generated';
-
-// eslint-disable-next-line import/no-cycle
 import useLiveItemBackup from '@src/client/backups/useLiveItemBackup';
 import formatLiveItem from '@src/client/formatters/formatLiveItem';
-
-// eslint-disable-next-line import/no-cycle
 import { extractItemId } from '@src/lib';
-
-// eslint-disable-next-line import/no-cycle
-import client from '..';
-
-import { useRpcBackup } from './useHealth';
+import client from '@src/client';
+import { useHealth } from '@src/client/hooks/useHealth';
 
 export default (tokenId: string | undefined, sellingTokenId: string | undefined) => {
     const graph = client.live.graph();
 
-    const backup = useRpcBackup();
     const updateToken = client.mutate.updateToken();
+
     useWatchLiveItemSubscription({
         client: graph,
         shouldResubscribe: true,
@@ -37,7 +30,7 @@ export default (tokenId: string | undefined, sellingTokenId: string | undefined)
         },
     });
 
-    useLiveItemBackup(backup, tokenId, sellingTokenId);
+    useLiveItemBackup(useHealth().graphProblem, tokenId, sellingTokenId);
 
     return null;
 };
